React code refactoring (#588)
* Add staging server

* Fix path

* Fix path

* Fix path

* Fix env path

* Fix build

* Run clone script from nantralPlatform

* Change env variable names

* Fix config

* React refactor (#587)

* Update 0005_migrate_data.py

* Add staging server (#579)

* Add staging server

* Fix path

* Fix path

* Fix path

* Fix env path

* Fix build

* Run clone script from nantralPlatform

* Change env variable names

* Fix config

* Add descriptions, format titles and unindex some pages (#582)

* Search engine optimization (#583)

* Add descriptions, format titles and unindex some pages

* Update map.html

* Update README.md

* Minor fixes (#584)

* Update 0005_migrate_data.py

* Update README.md

* :sparkles:  Add install script for UNIX systems (#585)

* :sparkles:  Add install script for UNIX systems

* :sparkles:  Add virtualenv install in script

* Add colors and verbose output

* :sparkles:  Configure environment variables in install (#586)

* :sparkles:  Configure environment variables in install

* :recycle:  Fix environment file name and location

* :recycle:  Fix file name

* Better code formatting

* Split housingMap code

* Split the code even more

* Split the code
